replace some tests, added tests for reset password
